[feat/#19] 사용자 지역 업데이트 기능 구현 완료

+++ src/main/java/com/app/weather/domain/region/domain/Region.java
@@ -24,9 +24,9 @@ public class Region extends BaseEntity {
 
     private String name;
 
-    private double lat;
+    private Float lat;
 
-    private double lon;
+    private Float lon;
 
     @OneToMany(mappedBy = "region",cascade = CascadeType.PERSIST)
     @Builder.Default
