+++ src/main/java/com/app/weather/domain/region/dto/request/LocationRequest.java
@@ -8,6 +8,6 @@ import lombok.*;
 @AllArgsConstructor
 public class LocationRequest {
 
-    private double userLat;
-    private double userLon;
+    private Float userLat;
+    private Float userLon;
 }
